feat: Adjusts storage dependencies
