+++ lib/feature_flagger.rb
@@ -11,7 +11,7 @@ require 'feature_flagger/feature'
 require 'feature_flagger/key_resolver'
 require 'feature_flagger/configuration'
 require 'feature_flagger/manager'
-require 'feature_flagger/feature_keys_migration'
+require 'feature_flagger/storage/feature_keys_migration'
 require 'feature_flagger/railtie'
 
 module FeatureFlagger
